requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.3
openai==1.107.2
openpyxl==3.1.5
pandas==2.3.2
pgvector==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
gunicorn==23.0.0
